projects: max31827-evkit : Modify platform from FTD2XX to linux
Since ftd2xx is no longer a platform, port the support to linux platform
and use ftd2xx as library instead.
