+++ projects/max31827-evkit/Makefile
@@ -1,5 +1,6 @@
 # Select the example you want to enable by choosing y for enabling and n for disabling
 EXAMPLE ?= basic
+LIBRARIES += ftd2xx
 
 include ../../tools/scripts/generic_variables.mk
