--- projects/max31827-evkit/src/platform/ftd2xx/platform_src.mk
@@ -1,10 +1,0 @@
-
-INCS += $(PLATFORM_DRIVERS)/$(PLATFORM)_i2c.h		\
-	$(PLATFORM_DRIVERS)/$(PLATFORM)_uart.h		\
-	$(PLATFORM_DRIVERS)/$(PLATFORM)_platform.h	\
-
-SRCS += $(PLATFORM_DRIVERS)/$(PLATFORM)_delay.c		\
-        $(PLATFORM_DRIVERS)/$(PLATFORM)_i2c.c		\
-	$(PLATFORM_DRIVERS)/$(PLATFORM)_uart.c		\
-	$(PLATFORM_DRIVERS)/$(PLATFORM)_platform.c	\
-
