+++ projects/max31827-evkit/src/platform/linux/main.c
@@ -1,6 +1,6 @@
 /*******************************************************************************
- *   @file   parameters.c
- *   @brief  Definition of FTD2XX platform data used by max31827 project.
+ *   @file   main.c
+ *   @brief  Main file for linux platform of max31827 project.
  *   @author Radu Sabau (radu.sabau@analog.com)
  *******************************************************************************
  * Copyright 2025(c) Analog Devices, Inc.
@@ -30,11 +30,14 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  ******************************************************************************/
-#include "parameters.h"
+#include "common_data.h"
 
-struct ftd2xx_i2c_init max31827_i2c_extra = {
-	.channel_config = false,
-	.extra_pins = true,
-	.extra_pins_dir = 0xFF,
-	.extra_pins_val = 0xBE
-};
+/******************************************************************************
+ * @brief Main function execution for FTD2XX platform.
+ *
+ * @return ret - Result of the enabled examples execution.
+ *****************************************************************************/
+int main()
+{
+	return example_main();
+}
