+++ projects/max31827-evkit/src/platform/linux/parameters.c
@@ -1,6 +1,6 @@
 /*******************************************************************************
- *   @file   parameters.h
- *   @brief  Definition of FTD2XX platform data used by max31827 project.
+ *   @file   parameters.c
+ *   @brief  Definition of linux platform data used by max31827 project.
  *   @author Radu Sabau (radu.sabau@analog.com)
  *******************************************************************************
  * Copyright 2025(c) Analog Devices, Inc.
@@ -30,25 +30,17 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  ******************************************************************************/
-#ifndef __PARAMETERS_H__
-#define __PARAMETERS_H__
+#include "parameters.h"
 
-#include "ftd2xx_uart.h"
-#include "ftd2xx_i2c.h"
-
-#ifdef IIO_SUPPORT
-#define INTC_DEVICE_ID	0
+#ifdef FTD2XX
+struct ftd2xx_i2c_init max31827_i2c_extra = {
+	.channel_config = false,
+	.extra_pins = true,
+	.extra_pins_dir = 0xFF,
+	.extra_pins_val = 0xBE
+};
+#else
+struct linux_i2c_init_param max31827_i2c_extra = {
+	.device_id = I2C_DEVICE_ID,
+};
 #endif
-#define UART_IRQ_ID	0
-#define UART_DEVICE_ID	0
-#define UART_BAUDRATE	0
-#define UART_EXTRA      NULL
-#define UART_OPS        &ftd2xx_uart_ops
-
-#define I2C_DEVICE_ID   0
-#define I2C_OPS         &ftd2xx_i2c_ops
-#define I2C_EXTRA       &max31827_i2c_extra
-
-extern struct ftd2xx_i2c_init max31827_i2c_extra;
-
-#endif /* __PARAMETERS_H__ */
