+++ projects/max31827-evkit/src/platform/linux/parameters.h
@@ -1,6 +1,6 @@
 /*******************************************************************************
- *   @file   main.c
- *   @brief  Main file for FTD2XX platform of max31827 project.
+ *   @file   parameters.h
+ *   @brief  Definition of linux platform data used by max31827 project.
  *   @author Radu Sabau (radu.sabau@analog.com)
  *******************************************************************************
  * Copyright 2025(c) Analog Devices, Inc.
@@ -30,14 +30,40 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  ******************************************************************************/
-#include "common_data.h"
+#ifndef __PARAMETERS_H__
+#define __PARAMETERS_H__
 
-/******************************************************************************
- * @brief Main function execution for FTD2XX platform.
- *
- * @return ret - Result of the enabled examples execution.
- *****************************************************************************/
-int main()
-{
-	return example_main();
-}
+#ifdef FTD2XX
+#include "ftd2xx_uart.h"
+#include "ftd2xx_i2c.h"
+#else
+#include "linux_i2c.h"
+#include "linux_uart.h"
+#endif
+
+#ifdef IIO_SUPPORT
+#define INTC_DEVICE_ID	0
+#endif
+#define UART_IRQ_ID	0
+#define UART_DEVICE_ID	0
+#define UART_BAUDRATE	0
+#define UART_EXTRA      NULL
+
+#define I2C_DEVICE_ID   0
+
+#ifdef FTD2XX
+#define I2C_OPS         &ftd2xx_i2c_ops
+#define UART_OPS        &ftd2xx_uart_ops
+#else
+#define I2C_OPS		&linux_i2c_ops
+#define UART_OPS	&linux_uart_ops
+#endif
+#define I2C_EXTRA       &max31827_i2c_extra
+
+#ifdef FTD2XX
+extern struct ftd2xx_i2c_init max31827_i2c_extra;
+#else
+extern struct linux_i2c_init_param max31827_i2c_extra;
+#endif
+
+#endif /* __PARAMETERS_H__ */
