+++ projects/max31827-evkit/src/platform/linux/platform_src.mk
@@ -1,0 +1,7 @@
+
+INCS += $(PLATFORM_DRIVERS)/$(PLATFORM)_i2c.h		\
+	$(PLATFORM_DRIVERS)/$(PLATFORM)_uart.h
+
+SRCS += $(PLATFORM_DRIVERS)/$(PLATFORM)_delay.c		\
+        $(PLATFORM_DRIVERS)/$(PLATFORM)_i2c.c		\
+	$(PLATFORM_DRIVERS)/$(PLATFORM)_uart.c
